% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmvss_mc.R
\name{pmvss_mc}
\alias{pmvss_mc}
\title{Monte Carlo Multivariate Subgaussian Stable Distribution}
\usage{
pmvss_mc(
  lower = rep(-Inf, d),
  upper = rep(Inf, d),
  alpha = 1,
  Q = NULL,
  delta = rep(0, d),
  which.stable = c("libstableR", "stabledist")[1],
  n = NULL
)
}
\arguments{
\item{lower}{the vector of lower limits of length n.}

\item{upper}{the vector of upper limits of length n.}

\item{alpha}{default to 1 (Cauchy). Must be 0<alpha<2}

\item{Q}{Shape matrix.  See Nolan (2013).}

\item{delta}{location vector.}

\item{which.stable}{defaults to "libstableR", other option is "stabledist".  Indicates which package
should provide the univariate stable distribution in this production distribution form of a univariate
stable and multivariate normal.}

\item{n}{number of random vectors to be drawn for Monte Carlo calculation.}
}
\description{
Computes probabilities of the multivariate subgaussian stable
distribution for arbitrary limits, alpha, and shape matrices
via Monte Carlo (thus the suffix \verb{_mc}).
}
\examples{

## print("mvpd (d=2, alpha=1.71):")
U <- c(1,1)
L <- -U
Q <- matrix(c(10,7.5,7.5,10),2)
mvpd::pmvss_mc(L, U, alpha=1.71, Q=Q, n=1e3)
mvpd::pmvss   (L, U, alpha=1.71, Q=Q)

## more accuracy = longer runtime
mvpd::pmvss_mc(L, U, alpha=1.71, Q=Q, n=1e4)

U <- c(1,1,1)
L <- -U
Q <- matrix(c(10,7.5,7.5,7.5,10,7.5,7.5,7.5,10),3)
## print("mvpd: (d=3, alpha=1.71):")
mvpd::pmvss_mc(L, U, alpha=1.71, Q=Q, n=1e3)


}
\references{
Nolan JP (2013), \emph{Multivariate elliptically contoured stable distributions:
theory and estimation}. Comput Stat (2013) 28:2067–2089
DOI 10.1007/s00180-013-0396-7
}
