% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkolm.R
\name{rkolm}
\alias{rkolm}
\title{Kolmogorov Distribution Random Variates}
\usage{
rkolm(n, nterms = 500)
}
\arguments{
\item{n}{}

\item{nterms}{}
}
\description{
Kolmogorov Distribution Random Variates
}
\examples{
rkolm(10)
}
